+++ CHANGES.rst
@@ -4,8 +4,8 @@ Products.MeetingLiege Changelog
 4.1rc10 (unreleased)
 --------------------
 
-- Nothing changed yet.
-
+- `TREASURY_GROUP_ID` suffixed Plone group `incopy` is now set in copy of items having finances advice in place of `observers` suffixed Plone group.
+  [gbastien]
 
 4.1rc9 (2020-05-08)
 -------------------